Fixed a few errors

+++ Math/stats.h
@@ -8,8 +8,8 @@ double binomial(int n, int x, float Pi);
 double binoSeries(int n, float Pi, int s, int e);
 double pow(double a, int x);
 
-long factorial(int n);
-long combination(int n, int r);
-long optiFact(int n, int r);
+unsigned int long long factorial(int n);
+unsigned int long long combination(int n, int r);
+unsigned int long long optiFact(int n, int r);
 
 #endif
